Fix: pmem_len and xram_len should be able to store 64*1024

+++ include/emu51.h
@@ -26,7 +26,7 @@ typedef struct emu51_callbacks emu51_callbacks;
 typedef struct emu51
 {
 	const uint8_t *pmem; /**< Read-only program memory */
-	uint16_t pmem_len; /**< Size of the @c pmem buffer,
+	int pmem_len; /**< Size of the @c pmem buffer,
 								must be power of 2 within 1k~64k */
 
 	/** Lower internal memory (address 0~127).
@@ -49,7 +49,7 @@ typedef struct emu51
 	uint8_t *sfr;
 
 	uint8_t *xram; /**< External memory, leave it NULL if not used */
-	uint16_t xram_len; /**< Size of the @c xram buffer,
+	int xram_len; /**< Size of the @c xram buffer,
 								must be power of 2 within 1k~64k */
 
 	uint16_t pc; /**< Program counter */
